feat(dashboard): add live chat functionality to server error page
